Fix bug with volume buttons for key event listener
